refactor(store-core): split activity-reducer into state + selectors (#6201)
activity-reducer.ts (503 LOC) mixed the mutation reducers and the read-only
selectors in one module. Split along the SRP seam the audit flagged:

- activity-state.ts     — state types + the pure reducers (createEmpty/
  applySnapshot/applyDelta/clearSession + the upsert/prune/shouldReplace core)
- activity-selectors.ts — read-only selectors (selectSessionEntries/
  selectActivityTree + the cross-session mission-control aggregate) + the
  ActivityTreeNode output type
- activity-internal.ts  — the shared null-prototype map helpers (emptyRecord/
  hasKey), intentionally NOT re-exported (stays private)
- activity-reducer.ts   — now a barrel: `export * from` both layers, so the
  historical `./activity-reducer` import path (index.ts + tests) is unchanged

Pure move — no logic touched. Function/type bodies were sliced verbatim from
the original. Public @chroxy/store-core surface is byte-identical (index.ts
unchanged). Enables tree-shaking the selector layer out of mutation-only
consumers.

Verified: store-core typecheck + full suite (1714) green; dashboard typecheck +
its activity dispatch test (14) green; store-core dist/ untouched (typecheck via
tsc --noEmit + vitest only, never bare tsc).
